requirements.txt
Flask>=2.0.0
bcrypt>=3.2.0
bdnd
autodl-api
cryptography>=3.4.8

